BAT: Change maximum number of channels from 2 to 64
Change maximum number of channels from 2 to 64.

Added a round-robbin frequency assignment from the command line so that
if you specify -F 100:200:250 on the command line, then all subsequent
channels will get assigned frequencies like this first set.  So by
specifying

    -c 6 -F 100:200:250

the resulting frequencies for the 6 channels will be the same as if you
specified

    -c 6 -F 100:200:250:100:200:250

Signed-off-by: Caleb Crome <caleb@crome.org>
Acked-by: Lu, Han <han.lu@intel.com>